pyproject.toml
[project]
name = "projeto-classificador-juridico"
version = "0.1.0"
description = ""
authors = [
    {name = "rick",email = "ricardoarieldesouzabatista0901@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11.9,<4.0"
dependencies = [
    "langchain (>=1.1.2,<2.0.0)",
    "openai (>=2.9.0,<3.0.0)",
    "fastapi (>=0.123.9,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
    "langchain-core (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "projeto_classificador_juridico", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
